Platform/Data_Governance_Orchestrator/pyproject.toml
[tool.poetry]
name = "Data_Governance_Cockpit"
version = "0.1.0"
description = ""
authors = ["HSKA"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0.1"
uvicorn = {extras = ["standard"], version = "^0.19.0"}
#pydantic = "^1.10.2"
pydantic = "^2.10.6"
#fastapi = "^0.85.1"
fastapi = "^0.115.8"
authlib = "^1.2.1"
python-keycloak = "^3.3.0"
icecream = "^2.1.3"
pymongo = "^4.8.0"
neo4j = "^5.21"
pyjwt = "^2.8.0"
httpx = "^0.27.0"
lazy = "^1.6"
python-multipart = "^0.0.20"
pymysql = "^1.1.1"
minio = "^7.2.16"
pandas = "^2.3.2"
requests = "^2.32.5"
python-jose = "3.5.0"

[tool.poetry.group.dev.dependencies]
schemathesis = "^3.19.7"
pytest-cov = "^4.0.0"
pytest-runner = "^6.0.0"
flake8 = "^5.0.4"
pylint = "^2.15.5"
bandit = "^1.7.4"
requests = "^2.28.1"
mongomock = "^4.1.2"
mypy = "^1.0.1"
types-pyyaml = "^6.0.12.6"
pytest = "^7.2.1"
isort = "^5.12.0"
pydocstyle = "^6.3.0"
grip = "^4.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
